refactor: change files which exposing components to PascalCase naming #38
