+++ src/common/components/index.ts
@@ -1,3 +1,3 @@
-export { GlobalOverray } from "./globalOverray";
-export { GlobalSpinner } from "./globalSpinner";
+export { GlobalOverray } from "./GlobalOverray";
+export { GlobalSpinner } from "./GlobalSpinner";
 export { GlobalHeader } from "./GlobalHeader";
