+++ src/gameLog/components/GameLogHeader.tsx
