+++ src/gameLog/components/GameLogTimeline.tsx
