+++ src/gomoku/components/UserSection.tsx
@@ -5,7 +5,7 @@ import Chip from '@mui/material/Chip';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 
-import { CurrentUser } from "./board";
+import { CurrentUser } from "./Board";
 import { useUsers } from "../context/usersProvider";
 
 export const UserSection: React.FC<{ currentUser: CurrentUser }> = ({
